generic Services install now works on Windows

+++ openmsistream/services/examples/runnable_example.py
@@ -1,4 +1,5 @@
 #imports
+import datetime
 from openmsistream.running import Runnable
 
 class RunnableExample(Runnable) :
@@ -16,7 +17,10 @@ class RunnableExample(Runnable) :
         parser = cls.get_argument_parser()
         args = parser.parse_args(args)
         test_file_name = 'runnable_example_service_test.txt'
-        with open(args.output_dir/test_file_name,'w') as fp :
-            msg = "This file was created to test running a generic Runnable's 'run_from_command_line' "
-            msg+= "method as a Service/daemon"
-            fp.write(msg)
+        if not (args.output_dir/test_file_name).is_file() :
+            with open(args.output_dir/test_file_name,'w') as fp :
+                msg = "This file was created to test running a generic Runnable's 'run_from_command_line' "
+                msg+= "method as a Service/daemon"
+                fp.write(msg)
+        with open(args.output_dir/test_file_name,'a') as fp :
+            fp.write(f'\nService rerun {(datetime.datetime.now()).strftime("on %Y-%m-%d at %H:%M:%S")}')
